Move Composition validation into its own file
We're hoping to move these out into webhooks soon.
