+++ internal/controller/apiextensions/composite/composed_test.go
@@ -43,130 +43,6 @@ import (
 	"github.com/crossplane/crossplane/internal/xcrd"
 )
 
-func TestRejectMixedTemplates(t *testing.T) {
-	cases := map[string]struct {
-		comp *v1.Composition
-		want error
-	}{
-		"Mixed": {
-			comp: &v1.Composition{
-				Spec: v1.CompositionSpec{
-					Resources: []v1.ComposedTemplate{
-						{
-							// Unnamed.
-						},
-						{
-							Name: pointer.StringPtr("cool"),
-						},
-					},
-				},
-			},
-			want: errors.New(errMixed),
-		},
-		"Anonymous": {
-			comp: &v1.Composition{
-				Spec: v1.CompositionSpec{
-					Resources: []v1.ComposedTemplate{
-						{
-							// Unnamed.
-						},
-						{
-							// Unnamed.
-						},
-					},
-				},
-			},
-			want: nil,
-		},
-		"Named": {
-			comp: &v1.Composition{
-				Spec: v1.CompositionSpec{
-					Resources: []v1.ComposedTemplate{
-						{
-							Name: pointer.StringPtr("cool"),
-						},
-						{
-							Name: pointer.StringPtr("cooler"),
-						},
-					},
-				},
-			},
-			want: nil,
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			got := RejectMixedTemplates(tc.comp)
-			if diff := cmp.Diff(tc.want, got, test.EquateErrors()); diff != "" {
-				t.Errorf("\nRejectMixedTemplates(...): -want, +got:\n%s", diff)
-			}
-		})
-	}
-}
-
-func TestRejectDuplicateNames(t *testing.T) {
-	cases := map[string]struct {
-		comp *v1.Composition
-		want error
-	}{
-		"Unique": {
-			comp: &v1.Composition{
-				Spec: v1.CompositionSpec{
-					Resources: []v1.ComposedTemplate{
-						{
-							Name: pointer.StringPtr("cool"),
-						},
-						{
-							Name: pointer.StringPtr("cooler"),
-						},
-					},
-				},
-			},
-			want: nil,
-		},
-		"Anonymous": {
-			comp: &v1.Composition{
-				Spec: v1.CompositionSpec{
-					Resources: []v1.ComposedTemplate{
-						{
-							// Unnamed.
-						},
-						{
-							// Unnamed.
-						},
-					},
-				},
-			},
-			want: nil,
-		},
-		"Duplicates": {
-			comp: &v1.Composition{
-				Spec: v1.CompositionSpec{
-					Resources: []v1.ComposedTemplate{
-						{
-							Name: pointer.StringPtr("cool"),
-						},
-						{
-							Name: pointer.StringPtr("cool"),
-						},
-					},
-				},
-			},
-			want: errors.New(errDuplicate),
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			got := RejectDuplicateNames(tc.comp)
-			if diff := cmp.Diff(tc.want, got, test.EquateErrors()); diff != "" {
-				t.Errorf("\nRejectDuplicateNames(...): -want, +got:\n%s", diff)
-			}
-		})
-	}
-}
-
 func TestRender(t *testing.T) {
 	ctrl := true
 	tmpl, _ := json.Marshal(&fake.Managed{})
